chore(schema): harden compatibility flow
Fix compatibility edge cases across core, app, and editor flows.

Replace stringly expectedType with enum-backed contract values.

Use Int for recent-usage window days to match Postgres make_interval.

Fail closed on compatibility checks and add safer save/preview parsing.

Tighten editor reactivity and reduce unsafe type assertions.

+++ modules/epistola-core/src/main/kotlin/app/epistola/suite/templates/validation/TemplateSchemaCompatibilityProperties.kt
@@ -5,7 +5,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 
 @ConfigurationProperties(prefix = "epistola.templates.schema-compatibility")
 data class TemplateSchemaCompatibilityProperties(
-    val recentUsageWindowDays: Long = 7,
+    val recentUsageWindowDays: Int = 7,
     val recentUsageSampleLimit: Int = 100,
     val statuses: Set<RequestStatus> = setOf(
         RequestStatus.PENDING,
